finished

+++ db/builder.cc
@@ -19,7 +19,7 @@ namespace leveldb {
 
 Status BuildTable(const std::string& dbname, Env* env, const Options& options,
                   TableCache* table_cache, Iterator* iter, FileMetaData* meta,
-                  VanillaBPlusTree<std::string, uint32_t>* btree) {
+                  VanillaBPlusTree<std::string, uint64_t>* btree) {
   Status s;
   meta->file_size = 0;
   iter->SeekToFirst();
